+++ client/reducers/index.js
@@ -1,67 +1,9 @@
-import { UPDATE_CLIENT_DETAILS, ADD_TOKEN_DETAILS, UPDATE_ORDER_ACTIVE_STATUS } from '../actions/clients'
-import { FETCH_ALL_CLIENTS_SUCCESS } from '../actions/admin'
-const initialUser = {
-}
+import { combineReducers } from 'redux'
 
-function updateClients (state = initialUser, action) {
-  // const details = action.details
-  switch (action.type) {
-    case UPDATE_ORDER_ACTIVE_STATUS:
-      return activeOrderHelper(state, action.orderStatus)
-    case UPDATE_CLIENT_DETAILS:
-      return updateClientHelper(state, action.details)
-    case ADD_TOKEN_DETAILS:
-      return authDetailHelper(state, action.details)
-    default:
-      return state
-  }
-}
+import updateClients from './clients'
+import admin from './admin'
 
-function activeOrderHelper (state, order) {
-  const { orderActive } = order
-  return {
-    ...state,
-    orderActive
-  }
-}
-
-function authDetailHelper (state, details) {
-  const { auth0Id, email, token } = details
-  return {
-    ...state,
-    auth0Id,
-    email,
-    token
-  }
-}
-
-function updateClientHelper (state, details) {
-  const { id, firstName, lastName, businessName, addressStreet, addressSuburb, addressCity, product, containers, price, orderActive, orderTimeStamp } = details
-  return {
-    ...state,
-    id,
-    firstName,
-    lastName,
-    businessName,
-    addressStreet,
-    addressSuburb,
-    addressCity,
-    product,
-    containers,
-    price,
-    orderActive,
-    orderTimeStamp
-  }
-}
-
-export function fetchAllClients (state = initialUser, action) {
-  switch (action.type) {
-    case FETCH_ALL_CLIENTS_SUCCESS:
-      return action.clients
-
-    default:
-      return state
-  }
-}
-
-export default updateClients
+export default combineReducers({
+  updateClients,
+  admin
+})
